requirements.txt
ipython==8.12.3
joblib==1.4.2
lmdb==1.7.5
lpips==0.1.4
matplotlib==3.8.4
numpy==2.3.5
pandas==2.3.3
Pillow>=9.3.0
plotly==6.5.0
pretrainedmodels==0.7.4
scipy==1.16.3
scikit-image
streamlit==1.52.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.5
certifi
